Rename project to python-piperci
