+++ piperci/gman/exceptions.py
