+++ tox.ini
@@ -30,7 +30,7 @@ max-complexity = 10
 
 [testenv:lint]
 deps = flake8>=3.6.0,<4
-commands = flake8 piedpiper tests
+commands = flake8 piperci tests
 skip_install = true
 usedevelop = false
 
@@ -39,4 +39,4 @@ usedevelop = true
 extras = test
 commands =
     pip freeze
-    pytest tests --cov=piedpiper --cov-report=term-missing --no-cov-on-fail {posargs}
+    pytest tests --cov=piperci --cov-report=term-missing --no-cov-on-fail {posargs}
